+++ src-tauri/src/commands/paste_clip.rs
@@ -1,10 +1,11 @@
+use crate::helpers::database::get_sqlite_db;
 use clipboard_win::raw::{close, open, set, set_without_clear};
 use clipboard_win::SysResult;
+use config::Config;
 use sqlx::Row;
 use std::ffi::c_void;
 use tauri::State;
 use tauri_plugin_sql::DbInstances;
-use tauri_plugin_sql::DbPool::Sqlite;
 use windows::Win32::Foundation::HWND;
 use windows::Win32::System::Threading::{AttachThreadInput, GetCurrentThreadId};
 use windows::Win32::UI::Input::KeyboardAndMouse::{
@@ -12,16 +13,16 @@ use windows::Win32::UI::Input::KeyboardAndMouse::{
 };
 use windows::Win32::UI::WindowsAndMessaging::{GetWindowThreadProcessId, SetForegroundWindow};
 
-use crate::constants::DB_PATH;
-
 #[tauri::command]
 pub async fn paste_clip(
+  config: State<'_, Config>,
   db_instances: State<'_, DbInstances>,
   clip_id: i64,
   target_window_ptr: usize,
 ) -> Result<(), String> {
-  let db_instances = db_instances.0.read().await;
-  let Sqlite(db) = db_instances.get(DB_PATH).unwrap();
+  let db = get_sqlite_db(config, db_instances)
+    .await
+    .map_err(|e| format!("Can't get SQLite DB: {}", e))?;
   let query_result = sqlx::query(
     "
     SELECT format_id, data
@@ -30,9 +31,9 @@ pub async fn paste_clip(
     ",
   )
   .bind(clip_id)
-  .fetch_all(db)
+  .fetch_all(&db)
   .await
-  .unwrap();
+  .map_err(|e| e.to_string())?;
 
   if let Err(e) = open() {
     return Err(format!("Can't open clipboard. {e}"));