+++ src-tauri/src/constants.rs
@@ -1,1 +1,0 @@
-pub const DB_PATH: &str = "sqlite:jcm-database.db";