+++ src/app/core/services/extended-dialog.service.ts
@@ -1,9 +1,11 @@
-import { Directive, Injectable, Type } from '@angular/core';
+import { Directive, inject, Injectable, Type } from '@angular/core';
 import { DialogService, DynamicDialogConfig, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { firstValueFrom } from 'rxjs';
+import { MonitoringService } from './monitoring.service';
 
 @Injectable()
 export class ExtendedDialogService extends DialogService {
+  private readonly monitoring = inject(MonitoringService);
 
   async openAsync<
     TComponent extends BaseDialog<TResult, TData>,
@@ -11,7 +13,7 @@ export class ExtendedDialogService extends DialogService {
     TData = TComponent extends BaseDialog<any, infer D> ? D : void
   >(
     component: Type<TComponent>, config: DynamicDialogConfig<TData>
-  ): Promise<TResult> {
+  ): Promise<TResult | null> {
     const mergedConfig: DynamicDialogConfig<TData> = {
       modal: true,
       focusOnShow: false,
@@ -19,8 +21,16 @@ export class ExtendedDialogService extends DialogService {
       ...config
     };
     const dialogRef = this.open(component, mergedConfig);
+    if (!dialogRef) {
+      this.monitoring.fatal('Can\'t get dialog reference for component: ' + component.name);
+      return null;
+    }
     const dialogComponent = await firstValueFrom(dialogRef.onChildComponentLoaded.asObservable());
     const dynamicDialogComponent = this.getInstance(dialogRef);
+    if (!dynamicDialogComponent) {
+      this.monitoring.fatal('Can\'t get dynamic dialog component for dialog component: ' + component.name);
+      return null;
+    }
     dialogComponent.onDialogLoaded(dynamicDialogComponent.data);
     return await firstValueFrom(dialogRef.onClose) as TResult;
   }