+++ release-notes.md
@@ -1,4 +1,4 @@
 # Fixes
 
-* Fixed autostart setting.
-* Fixed pinned clips panel behavior.
+* Fixed fraction digits appearance for "pinned clips" area size setting.
+* "HTML Format" support for Text plugin.